ADR1497-2021, change some fields' type from int to double to accommodate the floating numbers in the database

+++ cvpWebApi/Models/Reaction.cs
@@ -9,7 +9,7 @@ namespace cvpWebApi.Models
     {
         public Int64 reaction_id { get; set; }
         public int report_id { get; set; }
-        public int duration { get; set; }
+        public double duration { get; set; }
         public String duration_unit { get; set; }
         public String pt_name { get; set; }
         public String soc_name { get; set; }
